move item related test to item testing module
